requirements.txt
pytesseract==0.3.4
Flask==1.1.2
Image==1.5.32
Pillow==7.2.0
Flask-MySQLdb==0.2.0